% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_position_title_insert_oslo.R
\docType{data}
\name{nor_xxx_position_title_insert_oslo_b2020_insert_oslo_dt}
\alias{nor_xxx_position_title_insert_oslo_b2020_insert_oslo_dt}
\alias{nor_xxx_position_title_insert_oslo_b2019_insert_oslo_dt}
\alias{nor_xxx_position_title_insert_oslo_b2017_insert_oslo_dt}
\title{Position of a title for the Oslo insert}
\format{
\describe{
\item{long}{Location code.}
\item{lat}{Location name.}
}

An object of class \code{data.table} (inherits from \code{data.frame}) with 1 rows and 2 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 1 rows and 2 columns.
}
\usage{
nor_xxx_position_title_insert_oslo_b2020_insert_oslo_dt

nor_xxx_position_title_insert_oslo_b2019_insert_oslo_dt

nor_xxx_position_title_insert_oslo_b2017_insert_oslo_dt
}
\description{
Position of a title for the Oslo insert
}
\examples{
# 2020 borders
library(ggplot2)
q <- ggplot(mapping = aes(x = long, y = lat))
q <- q + geom_polygon(
  data = csmaps::nor_municip_map_b2020_insert_oslo_dt,
  mapping = aes(group = group),
  color = "black",
  fill = "white",
  size = 0.2
)
q <- q + annotate(
  "text",
  x = csmaps::nor_xxx_position_title_insert_oslo_b2020_insert_oslo_dt$long,
  y = csmaps::nor_xxx_position_title_insert_oslo_b2020_insert_oslo_dt$lat,
  label = "Oslo"
)
q <- q + theme_void()
q <- q + coord_quickmap()
q
}
\keyword{datasets}
