% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_municip.R
\docType{data}
\name{nor_municip_map_bxxxx_default_sf}
\alias{nor_municip_map_bxxxx_default_sf}
\alias{nor_municip_map_b2020_default_sf}
\alias{nor_municip_map_b2019_default_sf}
\title{Maps of Norwegian municipalities in sf format}
\format{
\describe{
\item{geometry}{Multipolygon}
\item{location_code}{Location code (municipality code).}
}

An object of class \code{sf} (inherits from \code{data.frame}) with 422 rows and 2 columns.
}
\usage{
nor_municip_map_b2020_default_sf

nor_municip_map_b2019_default_sf
}
\description{
This data is licensed under Creative Commons BY 4.0 (CC BY 4.0).
}
\details{
Borders for 2020 and 2019 are provided.
}
\keyword{datasets}
