% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_oslo_ward.R
\docType{data}
\name{oslo_ward_map_bxxxx_default_sf}
\alias{oslo_ward_map_bxxxx_default_sf}
\alias{oslo_ward_map_b2020_default_sf}
\title{Map of Oslo wards (bydeler) in sf format}
\format{
\describe{
\item{geometry}{Multipolygon}
\item{location_code}{Location code (municipality code).}
}
}
\usage{
oslo_ward_map_b2020_default_sf
}
\description{
This data is licensed under Creative Commons BY 4.0 (CC BY 4.0).
}
\details{
Borders provided for 2020.
}
\keyword{datasets}
